+++ examples/Examples/PIsData.hs
@@ -2,15 +2,20 @@
 
 module Examples.PIsData (tests) where
 
+import Data.Text.Encoding (encodeUtf8)
+
 import Test.Tasty
 import Test.Tasty.HUnit
 
 import Utils
 
-import Plutarch.Builtin (pforgetData)
+import Plutarch.Api.V1
+import Plutarch.Api.V1.Tuple (pbuiltinPairFromTuple, ptupleFromBuiltin)
+import Plutarch.Builtin (pforgetData, ppairDataBuiltin)
 import Plutarch.Lift (PLifted)
 import Plutarch.Prelude
 
+import Plutus.V1.Ledger.Credential
 import qualified PlutusTx
 
 --------------------------------------------------------------------------------
@@ -62,10 +67,10 @@ tests :: HasTester => TestTree
 tests =
   testGroup
     "Builtin PIsData instances"
-    [pBool, pInteger]
+    [pboolTests, pintegerTests, punitTests, ppairTests]
 
-pBool :: HasTester => TestTree
-pBool =
+pboolTests :: HasTester => TestTree
+pboolTests =
   testGroup
     "Builtin PIsData instances: PBool"
     [ testCase "pfromData (pdata True) ≡ True" $
@@ -82,8 +87,8 @@ pBool =
         testPDataCompat @PBool False
     ]
 
-pInteger :: HasTester => TestTree
-pInteger =
+pintegerTests :: HasTester => TestTree
+pintegerTests =
   testGroup
     "PInteger"
     [ testCase "pfromData (pdata 100) ≡ 100" $
@@ -92,4 +97,46 @@ pInteger =
         testPFromDataCompat @PInteger 100
     , testCase "PlutusTx.fromData (pdata 100) ≡ Just 100" $
         testPFromDataCompat @PInteger 100
+    ]
+
+punitTests :: HasTester => TestTree
+punitTests =
+  testGroup
+    "Builtin PIsData instances: PUnit"
+    [ testCase "pfromData (pdata ()) ≡ ()" $
+        testPToFrom @PUnit ()
+    , testCase "pfromData (PlutusTx.toData ()) ≡ ()" $
+        testPFromDataCompat @PUnit ()
+    , testCase "PlutusTx.fromData (pdata ()) ≡ Just ()" $
+        testPDataCompat @PUnit ()
+    ]
+
+ppairTests :: HasTester => TestTree
+ppairTests =
+  testGroup
+    "Builtin PIsData instances: PBuiltinPair & PTuple"
+    [ testCase "pfromData (pdata (I 1, B 0x41)) ≡ (I 1, I 2)" $
+        let x =
+              ppairDataBuiltin # pconstantData @PInteger 1
+                -- ByteString doesn't have a ToData instance - can't use pconstantData....
+                #$ pdata
+                $ pconstant $ encodeUtf8 "A"
+         in pfromData (pdata x)
+              `equal` x
+    , testCase
+        "pfromData (pdata (PTxId 0x41, PScriptCredential 0x82)) ≡ (PTxId 0x41, PScriptCredential 0x82)"
+        $ let x =
+                ppairDataBuiltin
+                  # pconstantData @PTxId "41" #$ pconstantData
+                  $ ScriptCredential "82"
+           in pfromData (pdata x) `equal` x
+    , testCase "ptuple isomorphism" $ do
+        let p =
+              ppairDataBuiltin
+                # pconstantData @PTxId "41" #$ pconstantData
+                $ ScriptCredential "82"
+            tup = pdata $ ptuple # pconstantData @PTxId "41" #$ pconstantData $ ScriptCredential "82"
+        pforgetData (pdata p) `equal` pforgetData tup
+        pfromData (pbuiltinPairFromTuple tup) `equal` p
+        ptupleFromBuiltin (pdata p) `equal` tup
     ]